fix server

+++ inference/server.py
@@ -52,7 +52,7 @@ def async_upload_photo(image, objects):
         objects: object to be segmented 
     segments are labeled: datetime_xmin_ymin_xmax_ymax.jpg
     """
-    image = np.array(image)
+    image = np.array(image, dtype = np.uint8)
 
     S3 = boto3.client('s3', aws_access_key_id = ACCESS_KEY_ID, aws_secret_access_key = SECRET_ACCESS_KEY)
